changed ofp type column to 0xff by default

+++ userland/examples_zc/zcount.c
@@ -242,7 +242,7 @@ struct ofp_match* ofpMatch;
 void process_ofp(struct ofp_header * ofp, char * output, char * type) {
     // printf("OFP: type(%u) xid(%u) - ", ofp->type, ofp->xid);
     
-    *type = ofp->type;
+    *type = (u_int16_t)ofp->type;
     
     switch (ofp->type) {
         case OFPT_PACKET_IN:
